imp: filter unwanted first-level regions via ISO3166-2
There are a few ugly regions (e.g. R2869330 or R3659532)
which are not actually ordinary administrations of the country
and thus should not be part of the hierarchy.
We try to exclude these using the ISO3166-2 code. The assumption
is that as soon as at least one administration has this code,
all valid administrations of this country also have the code.
Except for such ugly regions. Thus, we only take the regions
with the (correct) code.

In addition, the code was restructered a little and comments were
added.